Fri, Sep 30, 2022, 5:10 PM +00:00

+++ pset4/filter/helpers.c
@@ -9,7 +9,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j <= width; j++)
         {
-            average = round((image[i][j].rgbtGreen + image[i][j].rgbtRed + image[i][j].rgbtBlue) / 3);
+            average = (int)round((image[i][j].rgbtGreen + image[i][j].rgbtRed + image[i][j].rgbtBlue) / 3);
 
             image[i][j].rgbtGreen = average;
             image[i][j].rgbtBlue = average;
